feat(buffer): upgrade circular buffer to support 128KB+ sizes
Changed circular buffer implementation from uint16_t to uint32_t to
support buffers larger than 64KB. Updated SD card buffer to 128KB,
achieving ~290 KB/s for 500KB writes (up from ~263 KB/s).

Performance improvements:
- 250KB writes: 275 KB/s (+12 KB/s)
- 500KB writes: 290 KB/s (+27 KB/s)

Modified files:
- CircularBuffer: Changed size types from uint16_t to uint32_t
- SD Manager: Increased buffer from 60KB to 128KB
- WiFi/USB: Updated callback signatures for uint32_t
- CircularBuf_Safe: Updated wrapper functions

+++ firmware/src/Util/CircularBuffer.c
@@ -33,7 +33,7 @@
 #define BUF_START   cirbuf->buf_ptr
 #define BUF_END    (cirbuf->buf_ptr + cirbuf->buf_size - 1)
 
-void CircularBuf_Init(CircularBuf_t* cirbuf, int(*fp)(uint8_t*,uint16_t), uint16_t size)
+void CircularBuf_Init(CircularBuf_t* cirbuf, int(*fp)(uint8_t*,uint32_t), uint32_t size)
 {
     cirbuf->buf_ptr                = OSAL_Malloc(size);
     cirbuf->process_callback       = fp;
@@ -54,10 +54,10 @@ void CircularBuf_Init(CircularBuf_t* cirbuf, int(*fp)(uint8_t*,uint16_t), uint16
  *           while we are reading it, we read until we get the same answer twice in a row.  
  *           This method allows us to get a good reading without turning off interrupts to do it.
  *===========================================================================*/
-uint16_t CircularBuf_NumBytesAvailable(CircularBuf_t* cirbuf)
+uint32_t CircularBuf_NumBytesAvailable(CircularBuf_t* cirbuf)
 {
 
-    uint16_t num1,num2;
+    uint32_t num1,num2;
     do{
       num1 = cirbuf->totalBytes;
       num2 = cirbuf->totalBytes;
@@ -66,16 +66,16 @@ uint16_t CircularBuf_NumBytesAvailable(CircularBuf_t* cirbuf)
     return num1;
 }
 
-uint16_t CircularBuf_NumBytesFree(CircularBuf_t* cirbuf)
+uint32_t CircularBuf_NumBytesFree(CircularBuf_t* cirbuf)
 {
     return (cirbuf->buf_size - CircularBuf_NumBytesAvailable(cirbuf));
 }
 
 
-uint16_t CircularBuf_ProcessBytes(CircularBuf_t* cirbuf, uint8_t* bytesBuf, uint16_t maxBytes, int* error)
+uint32_t CircularBuf_ProcessBytes(CircularBuf_t* cirbuf, uint8_t* bytesBuf, uint32_t maxBytes, int* error)
 {
-    uint16_t bytesToSend  = 0;
-    uint16_t bytesRemoved = 0; 
+    uint32_t bytesToSend  = 0;
+    uint32_t bytesRemoved = 0; 
     *error = 0;
    
     bytesToSend = min(CircularBuf_NumBytesAvailable(cirbuf), maxBytes);
@@ -87,8 +87,8 @@ uint16_t CircularBuf_ProcessBytes(CircularBuf_t* cirbuf, uint8_t* bytesBuf, uint
             //start next transfer
             //make sure we don't transfer the out of bound data.
             if((cirbuf->removePtr + bytesToSend) > BUF_END){
-                uint16_t chunk1 = 0;
-                uint16_t chunk2 = 0;
+                uint32_t chunk1 = 0;
+                uint32_t chunk2 = 0;
 
                 // transfer the first part of data
                 chunk1 = (BUF_END - cirbuf->removePtr + 1);
@@ -145,7 +145,7 @@ uint16_t CircularBuf_ProcessBytes(CircularBuf_t* cirbuf, uint8_t* bytesBuf, uint
   CircularBuf_AddBytes
   Send out CF UDP Server message. Returns number of bytes sent.
   =====================================================================================*/
-uint16_t CircularBuf_AddBytes(CircularBuf_t* cirbuf, uint8_t* bytesBuf, uint16_t bytesToSend)
+uint32_t CircularBuf_AddBytes(CircularBuf_t* cirbuf, uint8_t* bytesBuf, uint32_t bytesToSend)
 {
     int numBytesCopied = 0;
     int wMaxPut;
